requirements.txt
streamlit
pandas
numpy
plotly
scipy
scikit-learn
opencv-python-headless
pytesseract
PyMuPDF
fpdf2
Pillow